move : Coupon Entity 분리
