+++ src/main/java/bc1/gream/domain/order/service/BuyService.java
@@ -22,8 +22,8 @@ import bc1.gream.domain.order.repository.OrderRepository;
 import bc1.gream.domain.order.repository.SellRepository;
 import bc1.gream.domain.product.entity.Product;
 import bc1.gream.domain.product.service.query.ProductService;
-import bc1.gream.domain.user.entity.Coupon;
-import bc1.gream.domain.user.entity.DiscountType;
+import bc1.gream.domain.user.coupon.entity.Coupon;
+import bc1.gream.domain.user.coupon.entity.DiscountType;
 import bc1.gream.domain.user.entity.User;
 import bc1.gream.domain.user.service.CouponService;
 import bc1.gream.global.exception.GlobalException;
